Simplify package code and files
