--- constants/queryOptions.js
@@ -1,7 +1,0 @@
-const QUERY_OPTIONS = {
-  rhyme: 'rel_rhy',
-  nearRhyme: 'rel_nry',
-  synonym: 'rel_syn'
-}
-
-module.exports = QUERY_OPTIONS
