+++ package.json
@@ -13,6 +13,7 @@
   "author": "Patrick El-Hage",
   "license": "ISC",
   "dependencies": {
+    "axios": "^0.18.0",
     "chalk": "^2.4.1",
     "commander": "^2.18.0"
   },
