--- src/util/createQueryString.js
@@ -1,5 +1,0 @@
-const querystring = require('querystring')
-
-const createQueryString = (query) => querystring.stringify(query)
-
-module.exports = createQueryString
